+++ vscode/microsoft-kiota/src/steps.ts
@@ -3,8 +3,9 @@ import * as vscode from 'vscode';
 import { Disposable, l10n, OpenDialogOptions, QuickInput, QuickInputButton, QuickInputButtons, QuickPickItem, Uri, window, workspace } from 'vscode';
 
 import { allGenerationLanguages, generationLanguageToString, KiotaSearchResultItem, LanguagesInformation, maturityLevelToString } from './kiotaInterop';
-import { findAppPackageDirectory, getWorkspaceJsonDirectory } from './util';
-import { isTemporaryDirectory } from './utilities/temporary-folder';
+import { findAppPackageDirectory, getWorkspaceJsonDirectory, isValidUrl } from './util';
+import { IntegrationParams, isDeeplinkEnabled } from './utilities/deep-linking';
+import { isFilePath, isTemporaryDirectory } from './utilities/temporary-folder';
 
 export async function filterSteps(existingFilter: string, filterCallback: (searchQuery: string) => void) {
     const state = {} as Partial<BaseStepsState>;
@@ -72,28 +73,17 @@ export async function searchSteps(searchCallBack: (searchQuery: string) => Thena
             validate: validateIsNotEmpty,
             shouldResume: shouldResume
         });
+        if (state.searchQuery && (isValidUrl(state.searchQuery) || isFilePath(state.searchQuery))) {
+            state.descriptionPath = state.searchQuery;
+            return;
+        }
         state.searchResults = await searchCallBack(state.searchQuery);
         if (state.searchResults && Object.keys(state.searchResults).length > 0) {
             return (input: MultiStepInput) => pickSearchResult(input, state);
         } else {
-            state.descriptionPath = state.searchQuery;
-            return (input: MultiStepInput) => inputPathOrUrl(input, state);
-        }
-    }
-
-    async function inputPathOrUrl(input: MultiStepInput, state: Partial<OpenState>) {
-        if (state.descriptionPath) {
+            vscode.window.showErrorMessage(l10n.t('No results found. Try pasting a path or url instead.'));
             return;
         }
-        state.descriptionPath = await input.showInputBox({
-            title,
-            step: step++,
-            totalSteps: 1,
-            value: state.descriptionPath || '',
-            prompt: l10n.t('Search or paste a path to an API description'),
-            validate: validateIsNotEmpty,
-            shouldResume: shouldResume
-        });
     }
 
     async function pickSearchResult(input: MultiStepInput, state: Partial<SearchState & OpenState>) {
@@ -107,7 +97,7 @@ export async function searchSteps(searchCallBack: (searchQuery: string) => Thena
         const pick = await input.showQuickPick({
             title,
             step: step++,
-            totalSteps: totalSteps,
+            totalSteps: totalSteps + 1,
             placeholder: l10n.t('Pick a search result'),
             items: items,
             shouldResume: shouldResume
@@ -118,7 +108,7 @@ export async function searchSteps(searchCallBack: (searchQuery: string) => Thena
     return state;
 }
 
-export async function generateSteps(existingConfiguration: Partial<GenerateState>, languagesInformation?: LanguagesInformation, isDeeplinkEnabled?: boolean) {
+export async function generateSteps(existingConfiguration: Partial<GenerateState>, languagesInformation?: LanguagesInformation, deepLinkParams?: Partial<IntegrationParams>) {
     const state = { ...existingConfiguration } as Partial<GenerateState>;
     if (existingConfiguration.generationType && existingConfiguration.clientClassName && existingConfiguration.clientNamespaceName && existingConfiguration.outputPath && existingConfiguration.language &&
         typeof existingConfiguration.generationType === 'string' && existingConfiguration.clientNamespaceName === 'string' && typeof existingConfiguration.outputPath === 'string' && typeof existingConfiguration.language === 'string' &&
@@ -126,13 +116,13 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
         return state;
     }
 
-    const isDeepLinkPluginNameProvided = isDeeplinkEnabled && state.pluginName;
-    const isDeepLinkGenerationTypeProvided = isDeeplinkEnabled && state.generationType;
-    const isDeepLinkPluginTypeProvided = isDeeplinkEnabled && state.pluginTypes;
-    const isDeepLinkLanguageProvided = isDeeplinkEnabled && state.language;
-    const isDeepLinkOutputPathProvided = isDeeplinkEnabled && state.outputPath;
-    const isDeepLinkClientClassNameProvided = isDeeplinkEnabled && state.clientClassName;
-
+    const deeplinkEnabled = deepLinkParams && isDeeplinkEnabled(deepLinkParams);
+    const isDeepLinkPluginNameProvided = deeplinkEnabled && state.pluginName;
+    const isDeepLinkGenerationTypeProvided = deeplinkEnabled && state.generationType;
+    const isDeepLinkPluginTypeProvided = deeplinkEnabled && state.pluginTypes;
+    const isDeepLinkLanguageProvided = deeplinkEnabled && state.language;
+    const isDeepLinkOutputPathProvided = deeplinkEnabled && state.outputPath;
+    const isDeepLinkClientClassNameProvided = deeplinkEnabled && state.clientClassName;
 
     if (typeof state.outputPath === 'string' && !isTemporaryDirectory(state.outputPath)) {
         state.outputPath = workspace.asRelativePath(state.outputPath);
@@ -143,6 +133,10 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
     const appPackagePath = findAppPackageDirectory(workspaceFolder);
     if (appPackagePath) {
         workspaceFolder = appPackagePath;
+    }
+
+    if (isDeepLinkOutputPathProvided && deepLinkParams.source && deepLinkParams.source.toLowerCase() === 'ttk') {
+        state.workingDirectory = state.outputPath as string;
     }
 
     let step = 1;
@@ -167,70 +161,13 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
                 return inputClientClassName;
             case 'plugin':
                 return inputPluginName;
-            case 'apimanifest':
-                return inputManifestName;
+            case 'other':
+                return chooseOtherGenerationType;
             case 'httpSnippet':
                 return inputHttpSnippetOutputPath;
             default:
                 throw new Error('unknown generation type');
         }
-    }
-    async function inputGenerationType(input: MultiStepInput, state: Partial<GenerateState>) {
-        if (!isDeepLinkGenerationTypeProvided) {
-            const items = [l10n.t('Generate an API client'), l10n.t('Generate a plugin'), l10n.t('Generate an API manifest'), l10n.t('Generate HTTP snippet')];
-            const option = await input.showQuickPick({
-                title: l10n.t('What do you want to generate?'),
-                step: step++,
-                totalSteps: 3,
-                placeholder: l10n.t('Select an option'),
-                items: items.map(x => ({ label: x })),
-                validate: validateIsNotEmpty,
-                shouldResume: shouldResume
-            });
-            if (option.label === l10n.t('Generate an API client')) {
-                state.generationType = "client";
-            }
-            else if (option.label === l10n.t('Generate a plugin')) {
-                state.generationType = "plugin";
-            }
-            else if (option.label === l10n.t('Generate an API manifest')) {
-                state.generationType = "apimanifest";
-            }
-            else if(option.label === l10n.t('Generate HTTP snippet')) {
-                state.generationType = "httpSnippet";
-            }
-        }
-        let nextStep = getNextStepForGenerationType(state.generationType?.toString() || '');
-        return (input: MultiStepInput) => nextStep(input, state);
-    }
-    async function inputClientClassName(input: MultiStepInput, state: Partial<GenerateState>) {
-        if (!isDeepLinkClientClassNameProvided) {
-            state.clientClassName = await input.showInputBox({
-                title: `${l10n.t('Create a new API client')} - ${l10n.t('class')}`,
-                step: step++,
-                totalSteps: 5,
-                value: state.clientClassName ?? '',
-                placeholder: 'ApiClient',
-                prompt: l10n.t('Choose a name for the client class'),
-                validate: validateIsNotEmpty,
-                shouldResume: shouldResume
-            });
-        }
-        updateWorkspaceFolder(state.clientClassName);
-        return (input: MultiStepInput) => inputClientNamespaceName(input, state);
-    }
-    async function inputClientNamespaceName(input: MultiStepInput, state: Partial<GenerateState>) {
-        state.clientNamespaceName = await input.showInputBox({
-            title: `${l10n.t('Create a new API client')} - ${l10n.t('namespace')}`,
-            step: step++,
-            totalSteps: 5,
-            value: typeof state.clientNamespaceName === 'string' ? state.clientNamespaceName : '',
-            placeholder: 'ApiSDK',
-            prompt: l10n.t('Choose a name for the client class namespace'),
-            validate: validateIsNotEmpty,
-            shouldResume: shouldResume
-        });
-        return (input: MultiStepInput) => inputOutputPath(input, state);
     }
     async function inputHttpSnippetOutputPath(input: MultiStepInput, state: Partial<GenerateState>) {
         while (true) {
@@ -268,6 +205,68 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
             state.outputPath = state.outputPath === '' ? 'output' : state.outputPath;
             break;
         }
+    }
+    async function inputGenerationType(input: MultiStepInput, state: Partial<GenerateState>) {
+        if (!isDeepLinkGenerationTypeProvided) {
+            const items = [
+                l10n.t('Client'),
+                l10n.t('Copilot plugin'),
+                l10n.t('HTTP snippet'),
+                l10n.t('Other')
+            ];
+            const option = await input.showQuickPick({
+                title: l10n.t('What do you want to generate?'),
+                step: step++,
+                totalSteps: 3,
+                placeholder: l10n.t('Select an option'),
+                items: items.map(x => ({ label: x })),
+                validate: validateIsNotEmpty,
+                shouldResume: shouldResume
+            });
+            if (option.label === l10n.t('Client')) {
+                state.generationType = "client";
+            }
+            else if (option.label === l10n.t('Copilot plugin')) {
+                state.generationType = "plugin";
+            }
+            else if (option.label === l10n.t('HTTP snippet')) {
+                state.generationType = "httpSnippet";
+            }
+            else if (option.label === l10n.t('Other')) {
+                state.generationType = "other";
+            }
+        }
+        let nextStep = getNextStepForGenerationType(state.generationType?.toString() || '');
+        return (input: MultiStepInput) => nextStep(input, state);
+    }
+    async function inputClientClassName(input: MultiStepInput, state: Partial<GenerateState>) {
+        if (!isDeepLinkClientClassNameProvided) {
+            state.clientClassName = await input.showInputBox({
+                title: `${l10n.t('Create a new API client')} - ${l10n.t('class')}`,
+                step: step++,
+                totalSteps: 5,
+                value: state.clientClassName ?? '',
+                placeholder: 'ApiClient',
+                prompt: l10n.t('Choose a name for the client class'),
+                validate: validateIsNotEmpty,
+                shouldResume: shouldResume
+            });
+        }
+        updateWorkspaceFolder(state.clientClassName);
+        return (input: MultiStepInput) => inputClientNamespaceName(input, state);
+    }
+    async function inputClientNamespaceName(input: MultiStepInput, state: Partial<GenerateState>) {
+        state.clientNamespaceName = await input.showInputBox({
+            title: `${l10n.t('Create a new API client')} - ${l10n.t('namespace')}`,
+            step: step++,
+            totalSteps: 5,
+            value: typeof state.clientNamespaceName === 'string' ? state.clientNamespaceName : '',
+            placeholder: 'ApiSDK',
+            prompt: l10n.t('Choose a name for the client class namespace'),
+            validate: validateIsNotEmpty,
+            shouldResume: shouldResume
+        });
+        return (input: MultiStepInput) => inputOutputPath(input, state);
     }
     async function inputOutputPath(input: MultiStepInput, state: Partial<GenerateState>) {
         while (true) {
@@ -335,7 +334,7 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
             state.pluginName = await input.showInputBox({
                 title: `${l10n.t('Create a new plugin')} - ${l10n.t('plugin name')}`,
                 step: step++,
-                totalSteps: 4,
+                totalSteps: 3,
                 value: state.pluginName ?? '',
                 placeholder: 'MyPlugin',
                 prompt: l10n.t('Choose a name for the plugin'),
@@ -343,14 +342,15 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
                 shouldResume: shouldResume
             });
         }
+        state.pluginTypes = ['ApiPlugin'];
         updateWorkspaceFolder(state.pluginName);
-        return (input: MultiStepInput) => inputPluginType(input, state);
+        return (input: MultiStepInput) => inputPluginOutputPath(input, state);
     }
-    async function inputPluginType(input: MultiStepInput, state: Partial<GenerateState>) {
+    async function chooseOtherGenerationType(input: MultiStepInput, state: Partial<GenerateState>) {
         if (!isDeepLinkPluginTypeProvided) {
-            const items = ['API Plugin', 'Open AI'].map(x => ({ label: x }) as QuickPickItem);
+            const items = ['API Manifest', 'Open AI Plugin'].map(x => ({ label: x }) as QuickPickItem);
             const pluginTypes = await input.showQuickPick({
-                title: l10n.t('Choose a plugin type'),
+                title: l10n.t('Choose a type'),
                 step: step++,
                 totalSteps: 4,
                 placeholder: l10n.t('Select an option'),
@@ -358,9 +358,13 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
                 validate: validateIsNotEmpty,
                 shouldResume: shouldResume
             });
-            pluginTypes.label === 'API Plugin' ? state.pluginTypes = ['ApiPlugin'] : state.pluginTypes = ['OpenAI'];
+            pluginTypes.label === 'API Manifest' ? state.pluginTypes = ['ApiManifest'] : state.pluginTypes = ['OpenAI'];
+            
         }
-        return (input: MultiStepInput) => inputPluginOutputPath(input, state);
+
+        Array.isArray(state.pluginTypes) && state.pluginTypes.includes('ApiManifest') ?
+            state.generationType = 'apimanifest' : state.generationType = 'plugin';
+        return (input: MultiStepInput) => inputOtherGenerationTypeName(input, state);
     }
     async function inputPluginOutputPath(input: MultiStepInput, state: Partial<GenerateState>) {
         while (!isDeepLinkOutputPathProvided) {
@@ -399,28 +403,31 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
             break;
         }
     }
-    async function inputManifestName(input: MultiStepInput, state: Partial<GenerateState>) {
+    async function inputOtherGenerationTypeName(input: MultiStepInput, state: Partial<GenerateState>) {
         if (!isDeepLinkPluginNameProvided) {
+            const isManifest = state.pluginTypes && Array.isArray(state.pluginTypes) && state.pluginTypes.includes('ApiManifest');
             state.pluginName = await input.showInputBox({
-                title: `${l10n.t('Create a new manifest')} - ${l10n.t('manifest name')}`,
+                title: `${isManifest ? l10n.t('Create a new manifest') : l10n.t('Create a new OpenAI plugin')} - ${l10n.t('output name')}`,
                 step: step++,
-                totalSteps: 3,
+                totalSteps: 4,
                 value: state.pluginName ?? '',
-                placeholder: 'MyManifest',
-                prompt: l10n.t('Choose a name for the manifest'),
+                placeholder: `${isManifest ? 'MyManifest' : 'MyOpenAIPlugin'}`,
+                prompt: `${isManifest ? l10n.t('Choose a name for the manifest') : l10n.t('Choose a name for the OpenAI plugin')}`,
                 validate: validateIsNotEmpty,
                 shouldResume: shouldResume
             });
         }
         updateWorkspaceFolder(state.pluginName);
-        return (input: MultiStepInput) => inputManifestOutputPath(input, state);
+        return (input: MultiStepInput) => inputOtherGenerationTypeOutputPath(input, state);
     }
-    async function inputManifestOutputPath(input: MultiStepInput, state: Partial<GenerateState>) {
+    async function inputOtherGenerationTypeOutputPath(input: MultiStepInput, state: Partial<GenerateState>) {
         while (true) {
+            const isManifest = state.pluginTypes && Array.isArray(state.pluginTypes) && state.pluginTypes.includes('ApiManifest');
+
             const selectedOption = await input.showQuickPick({
-                title: `${l10n.t('Create a new manifest')} - ${l10n.t('output directory')}`,
+                title: `${isManifest ? l10n.t('Create a new manifest') : l10n.t('Create a new OpenAI plugin')} - ${l10n.t('output directory')}`,
                 step: step++,
-                totalSteps: 3,
+                totalSteps: 4,
                 placeholder: l10n.t('Enter an output path relative to the root of the project'),
                 items: inputOptions,
                 shouldResume: shouldResume
@@ -462,6 +469,13 @@ export async function generateSteps(existingConfiguration: Partial<GenerateState
     return state;
 }
 
+function shouldResume() {
+    // Could show a notification with the option to resume.
+    return new Promise<boolean>((resolve, reject) => {
+        // noop
+    });
+}
+
 export async function generateHttpSnippetsSteps(existingConfiguration: Partial<GenerateHttpSnippetState>) {
     const state = {...existingConfiguration} as Partial<GenerateState>;
     const title = l10n.t('Generate HTTP snippet');
@@ -486,13 +500,6 @@ export async function generateHttpSnippetsSteps(existingConfiguration: Partial<G
 
     await MultiStepInput.run(input => inputOutputPath(input, state), () => step -= 2);
     return state;
-}
-
-function shouldResume() {
-    // Could show a notification with the option to resume.
-    return new Promise<boolean>((resolve, reject) => {
-        // noop
-    });
 }
 
 function validateIsNotEmpty(value: string) {