+++ vscode/microsoft-kiota/src/util.ts
@@ -163,3 +163,12 @@ export function allGenerationLanguagesToString(): string[] {
   let allSupportedLanguages: string[] = allGenerationLanguages.map(langEnum => KiotaGenerationLanguage[langEnum]);
   return allSupportedLanguages;
 }
+
+export function isValidUrl(url: string): boolean {
+  try {
+    new URL(url);
+    return true;
+  } catch {
+    return false;
+  }
+}